Changed struct field "trocas" to unsigned long int to avoid overflow

+++ header_files/sorts.h
@@ -8,7 +8,7 @@
 struct info_sort
 {
     double tempo;
-    int trocas;
+    unsigned long int trocas;
     int comparacoes;
 };
